Реорганизовала: переименовала Chat в Layout(в соотв.с дз 3), и выделила непосредственно чат в новый компонент Chat. Вынесла часть инлайновых стилей в классы. Добавила обёрток и убрала лишние.
