+++ src/components/Layout/Chat/Chat.scss
@@ -1,14 +1,14 @@
-@import '_variables.scss';
-@import 'ChatList/ChatList.scss';
+@import '../_variables.scss';
 
-main{
-  display: flex;
-  flex-direction: column;
+.chat-window{
+  height: 100%;
+  width: 100%;
+
+  overflow: auto;
 }
 
 footer{
   position: relative;
-  bottom: 0;
 
   .placeholder{
     display: none;
