+++ src/components/Layout/ChatList/ChatList.scss
@@ -1,4 +1,6 @@
 .chat-list{
     width: 200px;
     background-color: white;
+
+    overflow: auto;
 }
