+++ src/components/Layout/Layout.scss
@@ -1,0 +1,13 @@
+.layout{
+    height: 100vh;
+    
+    .main{
+        padding: 0;
+
+        display: flex;
+        flex-direction: row;
+        flex-grow: 1; 
+
+        overflow: auto
+    }
+}
