+++ src/components/Layout/Message/Message.scss
@@ -1,4 +1,4 @@
-@import '_variables.scss';
+@import '../_variables.scss';
 
 .message {
     width: $width-chat/2;
