+++ src/components/Layout/_variables.scss
