+++ src/index.scss
@@ -1,5 +1,7 @@
-@import 'components/Chat/Chat.scss';
-@import 'components/Chat/Message.scss';
+@import 'components/Layout/Layout.scss';
+@import 'components/Layout/ChatList/ChatList.scss';
+@import 'components/Layout/Chat/Chat.scss';
+@import 'components/Layout/Message/Message.scss';
 
 body {
   margin: 0;
